Fix warnings in malloc source files

+++ src/aligned_malloc.c
@@ -46,7 +46,7 @@ void* aligned_malloc(size_t align, size_t size)
 		 * We know we have to fit an offset value
 		 * We also allocate extra bytes to ensure we can meet the alignment
 		 */
-		uint32_t hdr_size = PTR_OFFSET_SZ + (align - 1);
+		size_t hdr_size = PTR_OFFSET_SZ + (align - 1);
 		void* p = malloc(size + hdr_size);
 
 		if(p)
